test: fix bugs found while auditing the last commit's scratch tests
042a728 ("feat: add tests") added 21 debug scratch files under src/test —
exploratory scripts with weak or zero assertions, not tests. This resolves
every one of them: extracted into real regression tests, or deleted where
the only content was perf/allocation exploration already covered by the
established Bench convention (see test-support.Bench).

Several were mid-investigation of real bugs that never landed a fix:

- Rect.inset: Int overflow in `2 * horizontal`/`2 * vertical` made a huge
  margin grow the rect instead of collapsing it (ScratchInsetSpec's own
  assertion failed against the code it shipped next to).
- Scrollbar.thumbRange: same overflow class — a large content/viewport pair
  wrapped negative and collapsed the thumb to 1 cell instead of sizing it
  proportionally.
- TableColumns.resolve: the equal-columns fallback sized a `Seq.fill` from
  an unclamped, caller-supplied columnSpan — Int.MaxValue in one cell could
  OOM the render thread. Now clamped at the one number that bounds anything
  worth allocating: the area's own width.
- FrameEncoder: a malformed Cell (its symbol carrying more than one
  grapheme cluster, e.g. a raw escape sequence smuggled in through a
  caller-supplied glyph parameter that bypasses Buffer.setString's
  per-cluster filter) reached the terminal verbatim. Added the guard at the
  encoder, the one choke point everything already routes through.
- Log.append(text: String) never split on embedded '\n', so a multi-line
  chunk (the ordinary shape of output piped from a process) rendered as one
  row with the newlines silently dropped.
- Locale bugs, the same unguarded `.toLowerCase`/`f"..."` pattern the
  project already guards correctly in core.KeyEvent and terminal.ColorDepth,
  found unguarded in: DataTable.filteredRows, DirectoryTree's sort key,
  SyntaxHighlighter.Language.of, Calendar's day-number formatting (which
  defeats the class's own stated purpose — Locale.ENGLISH exists so output
  doesn't depend on the machine it runs on), Chart.formatBound, the
  published runtime.formatDuration, and test-support.Bench's report table.
- DirectoryTree symlink traversal: a symlink inside `root` pointing outside
  it is followed, and nothing sandboxes it. Not changed here — that's a
  real behavior decision, not a mechanical fix — but the previously
  undocumented behavior is now spelled out on DirectoryTreeState, and
  DirectoryTreeSymlinkSpec pins the current contract so a change to it (in
  either direction) is deliberate rather than a silent drift.

Also adds ProgressSpec: core.Progress is CLAUDE.md's named single owner of
"where is this animation at elapsed" for every animated widget and timed
Effect, and had zero direct tests anywhere in the repo.

+++ widgets/src/main/scala/io/worxbend/tui/widgets/Scrollbar.scala
@@ -134,7 +134,12 @@ final case class Scrollbar(
     if trackLength == 0 then None
     else if contentLength <= visible then Option.when(thumbWhenFits)((0, trackLength))
     else
-      val size            = math.max(1, math.min(trackLength, trackLength * visible / contentLength))
+      // `trackLength * visible` in `Long`, then clamped: both factors are `Int`, and `trackLength` (a handful of
+      // terminal cells) times a `visible` in the tens of millions overflows `Int` and wraps negative, which
+      // `math.min(trackLength, ...)` then prefers over the correct, positive `trackLength` — collapsing the thumb to
+      // its 1-cell floor on content large enough to overflow instead of sizing it proportionally. Same fix as the
+      // `x + maxWidth` bound in `Buffer.setString`.
+      val size            = math.max(1, math.min(trackLength, (trackLength.toLong * visible / contentLength).toInt))
       val maxPosition     = contentLength - visible
       val clampedPosition = math.max(0, math.min(position, maxPosition))
       val start           = math.round(clampedPosition.toDouble / maxPosition * (trackLength - size)).toInt
